requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
eventlet>=0.35.2
gunicorn>=21.2.0
africastalking>=1.2.7
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
faster-whisper>=1.0.1
google-genai>=0.7.0
anthropic>=0.30.0
groq>=0.9.0